refactor: features/home 수정 및 적용
