+++ frontend/src/features/home/index.ts
@@ -1,0 +1,2 @@
+export { default as BestReviewItem } from './components/BestReviewItem';
+export { default as BestReviews } from './components/BestReviews';
